requirements.txt
python-telegram-bot==20.1
python-dotenv==1.0.0
aiohttp==3.8.5
